Fixed Map Size Restriction
Previously maps had to be a of a size that consisted of sides of the
power of two and both sides had to be equal. However this fix made it
such

The fix is crude.
If the size is slightly off then it will using a with/height of the next
biggest power of two for the generation process, however it will only
return a matrix of the restricted size.

Known Bugs:
Since the fix only returns a smaller section of the generated map, for
some width/height combinations the map can either loose resolution (In
terms of generation and not picture quality) or apear stretched (in both
matrix and picture).

A better fix would be revising the generation algorithm to not miss
calculate if sizes are not optimal